finish from agent

+++ tests/Feature/Cleaning/WorkerWalletDashboardConsistencyTest.php
@@ -86,13 +86,13 @@ it('returns the same worker financial values used by the Filament dashboard', fu
     $this->getJson('/api/v1/cleaning/worker/account/deposit')
         ->assertOk()
         ->assertJsonPath('workerId', $worker->id)
-        ->assertJsonPath('currentBalance', $expected['depositBalance'])
-        ->assertJsonPath('allowedDebtLimit', $expected['allowedDebtLimit'])
-        ->assertJsonPath('totalRevenue', $expected['totalRevenue'])
+        ->assertJsonPath('currentBalance', (int) $expected['depositBalance'])
+        ->assertJsonPath('allowedDebtLimit', (int) $expected['allowedDebtLimit'])
+        ->assertJsonPath('totalRevenue', (int) $expected['totalRevenue'])
         ->assertJsonPath('completedJobs', $expected['completedJobs'])
-        ->assertJsonPath('totalCommission', $expected['totalCommission'])
-        ->assertJsonPath('adminCommissionBalance', $expected['adminCommissionBalance'])
-        ->assertJsonPath('grossInvoicesAmount', round(
+        ->assertJsonPath('totalCommission', (int) $expected['totalCommission'])
+        ->assertJsonPath('adminCommissionBalance', (int) $expected['adminCommissionBalance'])
+        ->assertJsonPath('grossInvoicesAmount', (int) round(
             (float) $expected['totalRevenue'] + (float) $expected['totalCommission'],
             2,
         ));
